I no longer trust Sourceforge
Remember when Sourceforge got caught making the Gimp download for
Windows an adware-infested “download helper”?  Yes, I understand
that, for web sites to stay up, money has to be spent.  On the
other hand, it has been my understanding that Sourceforge’s
business model was to have annying ads during the download to pay
for hosting (and I remember when Sourceforge was asking ISPs to
host their files to help get open source software out there on the
Internet).

It was OK when the adware downloads were benefiting the developer
(see, for example, FileZilla).  It was not OK when Sourceforge took
over the Gimp for Windows project and replaced the download with
their adware infested crap.  When they got caught, Sourceforge gave
us some bullshit about how they were “maintaining abandoned
projects.”  Yeah, right.

This in mind, I will no longer upload MaraDNS software to
Sourceforge and will no link to the Sourceforge MaraDNS project
on the GitHub MaraDNS page.

I remember when uploading a MaraDNS to Sourceforge was a 20-minute
ordeal.  In the end, they finally made it only a 10-minute ordeal.
Downloading has always been a “look at ads for 30 seconds before we
finally give you the download” ordeal.  There’s a reason everyone
went over to GitHub.  On GitHub, an upload is a simple “git commit
-a -m 'I’m too lazy to write a real commit message' ; git push origin
master”.

More details on how Sourceforge broke their trust with users:

http://arstechnica.com/information-technology/2015/05/sourceforge-grabs-gimp-for-windows-account-wraps-installer-in-bundle-pushing-adware/