Another "no more Sourceforge" mention
In this mention, I refer to a 2002 MaraDNS download, which is on
Sourceforge (because that’s where people hosted their open source
back then) as a place where you can get a copy of MaraDNS’s first
GPG key, for people who want to establish that MaraDNS’s GPG key
is trustable.

I do refer to Sourceforge as the place to get historical MaraDNS
releases, or as a place which hosts other people’s files, but not
for anything current.