Deadwood tests: sqa_tcp_buffering passes again Ticket: https://github.com/samboy/MaraDNS/issues/48
Note that this is giving us some technical debt: By replacing "dig" with
"askmara-tcp", both the sqa_tcp_buffering and the sqa_bigpacket tests
are not properly testing the Deadwood feature in question.

But, in the short term, DNS-over-TCP is not Deadwood's strong point
and it's disabled by default.  I will open up a "Let's make
askmara-tcp better" ticket.