Add shell script to make a blacklist to block ad-based domains.
See https://github.com/samboy/MaraDNS/issues/69

To use this, look in deadwood-github/tools for the file
make.blacklist.sh.  Once in the directory with this
file:

This is a command which will use Steven Black's blacklist to add
a large blacklist to Deadwood.  To use this script, on a UNIX-like
system (e.g. Linux, maybe MacOS), type in the following command from
the directory containing this:

mkdir /etc/deadwood/
mkdir /etc/deadwood/execfile/
./make.blacklist.sh > /etc/deadwood/execfile/blacklist

Once this "blacklist" file is created, add the following line to
the end of one's dwood3rc file:

execfile("blacklist")

Note that each entry takes over one kilobyte to store, so Deadwood
needs about 70 megabytes to use this blacklist.