Update API to return dwood3rc dictionary element size to handle errors
If a dictionary dwood3rc element is not in use, we simply return 0
when its size is requested instead of a special "uninitialized" value.
This way, we can add the size of different dwood3rc dictionaries
without having to worry about unset dicts.

This relates to https://github.com/samboy/MaraDNS/issues/73