ip_blacklist is now ip_blocklist in dwood3rc files
I am keeping ip_blacklist because it’s important to not break
currently dwood3rc files.

Another article on the matter:
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6148600/

Again, I am not doing this to be politically correct.

Again, I do not endorse the extremes from the progressive left.  In
particular, I do not endorse "cancel culture".

Again, I am doing this in the interests of empathy and compassion.