Add "primary"/"replica" note to DNS Master document.
I will not change the wording--it was and still is the wording used by
other DNS servers and the wording used in the RFCs MaraDNS follows--but
I will add a note that the wording is outdated.

Now, excuse me while I link to another rant on this matter:

	http://antirez.com/news/122

It makes an excellent point, one I will repeat here:

	I believe that political correctness has a puritan root. As such
	it focuses on formalities, but actually it has a real root of
	prejudice against others. [...] I believe that this environment
	is making it impossible to have important conversations.

That said, the terms are outdated.  They can bring up both racial and
sexual connotations which needlessly distract from concentrating on
running code and administering servers.