Adding Lua 5.1 fork (lunacy) and mmLunacyDNS to MaraDNS tree
This fork of Lua 5.1 aims to be fully compatible with Lua 5.1
scripts.  The reason for this fork is so I can add some extra
stuff to the "lunacy" scripting binary, namely a version of bit32
and the luafilesystem suite.

The reason for Lua 5.1 is so, if performance is ever an issue, we
can replace lunacy with LuaJIT.