coLunacyDNS: Begin work implementing coroutines
The processQuery process is now a Lua coroutine.  Since we do not
have the global space, coroutine.yield has the name coDNS.solve here
(as in, solve a DNS query, which is where we need to do the coroutine
magic).

I have also added 16-bit random numbers, since that is what DNS usually
needs.