Changing default UID/GID from 90909/90910 to 707
Why 707:

707 is easy to remember.

707 is not used over at https://svnweb.freebsd.org/ports/head/UIDs?view=co

707 is high enough it will not be a dynamically allocated UID as per
https://www.debian.org/doc/debian-policy/ch-opersys.html#uid-and-gid-classes
System UIDs are allocated like this: Start at 100, go up one like 101,
102, 103, 104, etc. as needed for system accounts.  Right now, there
are about 300 of those as per
https://fedoraproject.org/wiki/Features/1000SystemAccounts from 2011.
Note: The Debian document claims that 60000-64999 are allocated in some
master list somewhere, but I have been unable to locate this supposed
master list.  So, between 100 and 999 it is.

The Linux standard says system accounts stop at 500:
https://refspecs.linuxfoundation.org/LSB_5.0.0/LSB-Core-generic/LSB-Core-generic/uidrange.html
*HOWEVER* FreeBSD, Debian, and Fedora/RedHat/CentOS start user accounts
at UID 1000, **NOT** 500.  See the fedoraproject Wiki page above for
discussion of this.

Anything above 65536 is fair game for systems with large number of
users, and should not be used for system accounts.  It would had been
nice if they had reserved a larger block once we had 32-bit UIDs, like
65537-16777216 for system accounts, but that did not happen.