More doc changes for new 707 default UID for Duende.
Some more notes:

* CentOS 7 in /etc/login.defs starts new users at 1000
* Ditto with CentOS 8
* Old versions of RedHat/CentOS did start at 500, as per outdated notes at
  https://wiki.cam.ac.uk/cl-sys-admin/UID/GID_allocation

Now, the correct way to do this is to have a script create a "maradns"
user, then have all of the C code use a user name, **not** user ID for
IDs used by the sandbox, but I haven’t implemented that yet.

Also, the UIDs are only used for sandbox protection; anything non-root
is usable for this purpose, even if the ID happens to belong to someone,
this is not a significant security risk:  MaraDNS/Deadwood/coLunacyDNS do
not write to any files, except the cache file, which can be disabled.

Even when we use a cache file, we can protect things by having root own
the /etc/deadwood dir, have Deadwood’s chroot be /etc/deadwood/sub/,
write the cache to /etc/deadwood/sub/cache/cache_file, and give
/etc/deadwood 700 perms.  Deadwood is root when it does the chdir(),
so we can enter a directory only root can see, and have it owned by
any other user, who will not be able to access the directory unless a
root-owned process enters it first.