mararc.ej: Update UID to 707
This is part of my project to update the sandboxing.

Next: Have zone files and mararc in /etc/maradns/config

chmod 700 /etc/maradns
chmod 700 /etc/deadwood

MaraDNS needs to be able to read zone files *after* becoming
a non-root user, so while mararc is read as root, the zone
files need to be in a directory MaraDNS can read as the 707/99
user.

Deadwood needs to be able to read and write its cache as a non-root
user.

But, this will be done later this week.  The attack surface is quite
small, and no one has noticed it over the years until I did while
updating the MaraDNS UID.