coLunacyDNS: We now have an API for reading a file
Example use of this new API:

-------------------------------------------------------------------------
if not coDNS.open1("filename.txt") then
  return {co1Type = "serverFail"}
end
line = ""
while line do
  if line then coDNS.log("Line: " .. line) end
  line = coDNS.read1()
end
-------------------------------------------------------------------------

The calls are: coDNS.open1(filename), coDNS.read1(), and coDNS.close1()

Only a single file can be open at a time.  If coDNS.open1() is called
when a file is open, the currently open file is closed before we attempt
to open the new file.  If coDNS.solve() is called while a file is open,
the file is closed before we attempt to solve the DNS query.  If we exit
processQuery() while a file is open, the file is closed as we exit the
function.  Files are also closed when we finish parsing the Lua
configuration file used by coLunacyDNS, before listening to DNS queries.

The filename must start with an ASCII letter, number, or the _
(underscore) character.  The filename may contain only ASCII letters,
numbers, instances of "." (the dot character), or the _ character.
In particular, the filename may not contain /, \, or any commonly
used directory separator.

The file has to be in the same directory that coLunacyDNS is solved
from.  The file may only be read; writing to the file is not possible.

coDNS.read1() reads a single line from the file.  Any newline is
stripped from the end (unlike Perl, coLunacyDNS does not require a chop);
NUL characters in the line also truncate the string read.  If a line
is read from the file, coDNS.read1() returns the line which was read.
Otherwise, coDNS.read1() returns the "false" Lua boolean value.

coDNS.close1() closes an open file; a file is also closed when
opening another file, ending processQuery(), or calling coDNS.solve().
It is mainly here to give programmers trained to close open files
a function which does so.