Lunacy: Use SipHash to hash strings
See http://lua-users.org/wiki/HashDos for discussion

Look at this URL:

http://lua-users.org/files/wiki_insecure/power_patches/5.1/lua_5.1_second_hash_fix.patch

This solves the problem, but is far too complicated for my tastes.
There may be a performance penalty, but for my purposes, I want every
single last string hash to be calculated with SipHash.  This version
of Lunacy is used in network code, where a hash collision is dangerous.

Next: Securely seed the SipHash key.