lua.c (Lunacy): Seed SipHash key with random numbers
Some discussion here: Lua, like a lot of other programming languages,
is vulnerable to attacks where an attacker can make multiple strings
which hash to the same value.  Lua 5.2 has some mitigation against
these attacks, but Lua 5.1 (the base for Lunacy) doesn't.

This in mind, I have patched Lunacy to use SipHash as the core hash
compression function.  I have also, for the benefit of the Lua
community, made versions of these patches availale for both Lua 5.1
and Lua 5.4.

In my performance testing, I can not see a real-world speed difference
between using Lua's hash compressor and using 64-bit SipHash 2-4 on
a either a i7-7600U from 2017 nor a Core Duo T8100 from 2008 when
running code in 64-bit mode.

There is a 20% slowdown with real-world code when we compile Lunacy as
an i386 binary compiled with GCC 3.2.3 from 2002.  Personally,
i386 is rare enough these days that I am willing to live with a 20%
slowdown to keep the code base unified and secure, but there
may be ways to reduce the slowdown, such as using a SSE2 version
of SipHash [1] or using HalfSipHash 1-3 [2].

[1] https://github.com/floodyberry/siphash

[2] halfSipHash is described at https://github.com/veorq/SipHash
    SipHash 2-4 (standard SipHash) means we run two rounds of the
    core "compression" function for each 8 bytes (4 bytes in
    halfSipHash) of input, and four rounds of the "compression"
    function at the end of input.  With SipHash 1-3, we run only
    one round when processing input, and three rounds after
    processing input.