all tests: Add testVars so we can change distro-specific stuff quickly
For example, the AWK script in CentOS 8 to take the output of "ps auxw"
and print just the process ID and process name is '{print $2 " " $11}'
but there’s a pretty big chance this will change the next time they
mess around with procps.  So, we have one source of truth for that
AWK command, so we only have to change one line when ps changes.