coLunacyDNS: Found out what the gcov issue was
If a process is terminated, the gcov information is not updated.

So, I have created a custom SIGALRM (*not* SIGTERM, because processes
which do not terminate upon SIGTERM are acting very rudely) which will,
when -DGCOV is set, terminate the process by telling the select() loop
to stop (code which is already in place because Windows service code
does something similar to stop a service).  This way, we get to the
final line of main() and terminate normally, which allows Gcov to
update its coverage numbers.

SIGALRM will, in normal mode, inform the user that coLunacyDNS does
not reload configuration files when it gets an ALRM signal, and to
just restart the server.