coLunacyDNS: 72.22% Test coverage
IP6 tests updated to reach all lines in IPv6 parser except one.
Line 314 will *never* be reached but I am keeping it to keep the
IPv6 parser robust.

Lines 422-424 will not be covered: I am not going to remove
/dev/urandom system wide for a SQA test (maybe a docker-only one)
(But these lines are quite useful when I compile a win32 service
 with the wrong flags)

Line 462 can not be reached in any of this code, but I am keeping
it around so *that* library remains robust (I *could* reach this
line by changing the C code to call rand32() *before* init_rng(), but
nah)

So that takes us to line 536, which I believe I can get to by
leaving returnIp blank in the Lua return table in processQuery().