Deadwood SQA: Disable sqa_tcp_buffering
I am rather concerned that these tests, which work like a charm
in CentOS 8 + Podman, are failing in Ubuntu 20.04 + Docker.  It
*probably* indicates a bug with the Linux kernel or the libc TCP
libraries, since these tests have been running just fine for over
a decade.

I do not have the resources right now to deal with this bug.
DNS-over-TCP in Deadwood is only there for strict RFC compliance,
and I should probably disable DNS-over-TCP in Deadwood until
when and if I have time to deal with these corner case bugs.

https://github.com/samboy/MaraDNS/issues/91