Update API of dw_get_dname to allow return of label count
dw_get_dname now allows the number of labels to be returned to
the called via a pointer.  Number of labels is how many levels
of subdomains we have, e.g.

a.kabah.foo has three labels
a.b.kabah.foo has four labels
a.b.c.kabah.foo has five labels

and so on.