commit a0829666efa2a5dd20cb5322403566d88a0be603 passes all tests
Looks like the new Ubuntu 22.04 kernel finally fixed the
“set argv[0]/argv[1]/etc. to NULL” trick to make processes
seg fault.  Because of this, the userspace test to make
sure coLunacyDNS can handle argv[] with NULL values has different
output.  The test finally has been updated, and all is good again.

This is the first step in updating MaraDNS for Ubuntu 22.04.  The
next step is to update the Docker container for testing to run
Ubuntu 22.04 in userspace (instead of Ubuntu 20.04).  Then we
have to update the Deadwood makefile filename (makefile.ubuntu2204
instead of makefile.ubuntu2004), the docs, etc. to point out the
supported version is now Ubuntu 22.04.

We will have to repeat this process again every two years, until
I am no longer updating MaraDNS.

(Null edit to CHANGELOG.TXT)