gcc 4.4 test doesn’t make as much sense in 2022
It has been a “gcc compile”, *not* gcc 4.4 compile test for a long
time.  There may be value in running a test for the oldest version
of gcc available as an Ubuntu 22.04 package, but that would come
later.

Right now, let’s just get all of the tests to pass again by disabling
this test.