New test: sqa_cname_bad_bailiwick
OK, let’s have this when querying the example.com name server:

test.example.com. CNAME test.example.net.
test.example.net. A 10.5.5.5

Then let’s have this at the example.net name server:

test.example.net. A 10.7.7.7

Right now, Deadwood resolves test.example.net as having the IP 10.5.5.5
because there’s a way to accept out-of-bailiwick CNAME glue records
without corrupting the cache, and MaraDNS resolved things this way
back in 2001 because that was how BIND at the time handled it (mind you,
BIND would allow the cache to get corrupted, but it was a long time
ago).

However, BIND has long since fixed this bug, so Deadwood should handle
this the same way BIND does: It should discard the out of bailiwick CNAME
glue record and restart asking for test.example.net.

To define the problem, the first step is to make a test showing the
problem.

The next step is to fix the problem.

Then we add the test to the standard battery of regression tests for
Deadwood to make sure we don’t break this in the future.

Also: I fixed the year in 00README.FIRST