OK, Deadwood finally discards out of bailiwick CNAME answers
Deadwood used to allow out of bailiwick answers in a CNAME chain,
since:

* They are not a security threat
* They sometimes allow answers to be solved more quickly

However, let us suppose we have this:

blog.foo. CNAME blog.new-isp.foo.
blog.new-isp.foo. A 10.1.2.3

But the real answer for blog.new-isp.foo is 10.9.8.7, then
Deadwood would had stored the wrong IP for the blog.  This isn’t
a security problem, but it’s an issue because most DNS servers
would discard the CNAME glue while Deadwood accepted it.

This bug was filed in 2015.  I said, at the time, that I could
not fix it because I was, at the time, a single parent working a
full time job.  Now that I am semi-retired after failing to find
an employer in the tech sector emphatheic to the needs of a single
parent, I have time to fix this.