dwood2rc_s_ipv4_bind_addresses test: Remove perl
I am slowly going through the automated tests and replacing any usage
of perl with Lua instead.  This way, we do not need perl to run the
tests.  Mind you, we use other non-POSIX stuff like the -P option
to netstat to run this test, but it’s progress.

https://www.unix.com/man-page/posix/1M/netstat/

Since busybox has netstat, I’ll consider this test done with it
works with either busybox netstat or normal netstat.