RANT: There is no reason to break egrep and fgrep
This commit is an example of the amount of headache just one open
source project has to go through replacing egrep and fgrep.  fgrep has
been around for at least 25 years and egrep appears to have been around
for a similar period of time.

The offending diffs:

https://git.savannah.gnu.org/cgit/grep.git/commit/?id=a9515624709865d480e3142fd959bccd1c9372d1

https://git.savannah.gnu.org/cgit/grep.git/commit/?id=216f754287f2123f45d274f0a003182524efd43d

There is no harm in having the following shell script for egrep:

exec grep -E "$@"

And the following shell script for fgrep:

exec grep -F "$@"

The amount of effort I had to make to replace egrep with grep -E
eclipses the amount of effort of having two tiny shell scripts there.

Multiply this effort by the number of other scripts out there.

When one has deployed code which is on millions of computers, one should
not change that deployed code without a very very good reason.

Related discussion:

https://news.ycombinator.com/item?id=33189503