Remove mararc_n_notthere_ip test
This is for a feature someone paid me a few dollars to implement
back in early 2009 for a now long-gone startup company.

It doesn’t work with the authoritative only MaraDNS and it was never
implemented for Deadwood.

Today, you would have to pay me a five figure sum to implement
notthere_ip for Deadwood.