Dockefile: I like to have a full POSIX-like system in Docker images
There’s a debate I have had at jobs where it was debatable whether
a Docker image should have a full POSIX-like system via Busybox.  I
say yes, it’s worth making the Docker image 1,774,040 bytes bigger
to have all of POSIX etc. there (e.g. vi and so on) in case I need
to go in to the image and work on stuff.

Also, in a recent commit I have removed some cruft: An outdated
version of output.success.5.3 for a 10-year-old version of CentOS
that was in sqa_valgrind