sqa_valgrind: I found the Ubuntu 22.04 procps bug
The problem is this:

ps -e doesn’t show the name “valgrind” with procps.  It shows a name
like “memcheck-”.  The way I work around this is to look for valgrind with
“ps -e” (Busybox will be able to find the process this way).  If that
doesn’t work, we then use “ps auxw”, which doesn’t work in Busybox and
returns an error in Toybox, but will find the process with ubuntu procps.