commit 82cab01dd9a1b82854642fd76a479b5f85cdb95e passes all tests
Also, Deadwood and MaraDNS 3.4.04 released.  This is the legacy branch.

As an aside, Deadwood no longer passes SQA tests; the tests are OK but
the OS running the tests has changed and I would have to dig up an old
CentOS 6 or CentOS 8 image to get them to pass.  The tests 80-90% pass
and eyeballing some of the failing tests looks like the failures are
caused by OS changes.

My plan for the next 3.4.XX release of MaraDNS is to replace the
testing files with 3.5.XXXX tests, removing any tests for 3.5.XXXX
specific features or updates.  Then, add a Dockerfile for the testing
system, which I think I will make Alpine 3.14 for the older branch.

I might also make an Oracle VirtualBox image with the testing OS for
MaraDNS 3.4.  This way, I will have a long-term testing setup.