Begin work on getting all tests to pass in MaraDNS 3.4 branch again
The goal here is to have it so Deadwood passes all tests in Alpine 3.14

The reason I choose Alpine 3.14 is because it's a small Linux.  It's,
more to the point, small enough to fit on DVD including all packages
in the "main" directory (the testing suite for MaraDNS/Deadwood uses
only main packages).  This way, 5-10 years from now, even when my
Alpine Docker image is long gone and Alpine 3.14 is no longer online,
I will still be able to make a testing image for MaraDNS/Deadwood 3.4.
I have made a download of both the x86_64 and aarch64 versions of Alpine,
in case aarch64 overcomes x86_64 to the point new computers are Arm64
instead of Intel x86 machines -- this has already happened with Mac
computers.