Begin project to rewrite “ej” tools in Lua
One of the goals I have with MaraDNS is that no part of the
code will be dependent on a single implementation to run.  So far, I
have done the following:

* I have updated all of the tests to no longer use perl, since there
  is only one really standard implementation of perl.
* I have updated all of the tests to use whatever subset and/or superset
  of POSIX both Ubuntu 22.04 and Alpine Linux support.  The biggest
  offender is `ps`, and the culprit here looks to be Busybox’s very
  limited and yes buggy implementation of `ps`, but I have workarounds
  for it.

There are multiple implemtations of Lua 5.1:

* PUC Lua 5.1 (the original one) and my fork of it, Lunacy
* LuaJIT
* Gopher-Lua in Go

While there are multiple implemtations of Python, I will *not* use
Python as part of the MaraDNS toolchain.  The reason is this: With
Python 3, they threw out a long established standard (Python 2) and
replaced it with something similar, but incompatible.  I will not
depend on an ecosystem with a history of breaking their own standards
and compatibility.

It is true that 5.2 in some ways breaks 5.1, 5.3 breaks 5.2, and so on,
but with Lua, there are multiple independent implementations of Lua
5.1 and it’s a frozen, established standard.  And, unlike Perl and
unlike Python, it’s tiny and trivial to install if one has a C compiler,
and an implementation of it is part of MaraDNS (for coLunacyDNS).