ej2html.new: Other suggestions for dealing with Y2038
Here in the 2020s, we have 32-bit X86 Linux distros with a real
64-bit time_t (Alpine Linux, etc.).  We have a non-portal Linux
syscall which can get a 64-bit timestamp even if all of libc() is
still in the 32-bit time_t dark ages.  We have an entire repo which
can do a full localtime() after Y2038 even if time_t remains 32-bit.

There is no reasonable excuse to deploy a new Linux system in the 2020s
which will fail on January 19, 2038.