ej2txt.new: More regex massaging
The reason why we don’t really have a formal definition of EJ
files is because the regexes are adjusted until the end documents
look nice.  The whole EJ system is just a bunch of regex hacks
which works pretty well for having one document which is a text
file, web page, and man page.  And for having a man page without
having to dive in to *ROFF.