ej2txt.new: Finish up, space before/after <pre>
I think ej2txt.new is done.  I’ll need to test it, but it makes
nice looking text, and it solves the old issue where text files
would be completely rewritten if using a system with Links when
the previous system only had Lynx.

This way, the text files should only change when their underlying
text is changed.