The Lua EJ scripts have a UTF-8 problem.
The fmt() call, which fits text so it doesn’t wrap on an 80-column
terminal, assumes that each character is an ASCII character
when determining how long a line is.

However, if there is UTF-8, each character will be calculated as
being more than one unit wide.  With languages which use mostly
ASCII (e.g. Spanish), things will be mostly OK, since the only
effect will be that some lines with accents will be shorter than
expected.

In other languages with little ASCII (Greek, Cyrillic, etc.) the
current form of fmt() completely breaks down.

My plan is to look for high-bit characters and count each UTF-8
code point as being precisely one unit wide.  This breaks down in
languages with a lot of accents (Vietnamese, Thai, etc.) and even
breaks down in CJK (stuff like 你好 is twice as wide per code point
compared to ASCII letters) but properly supporting those kinds of
languages goes beyond the scope of a tiny language like Lua5.1/Lunacy.
People using those languages will need to just be content with the
translated HTML version of the documentation or create a Lunacy fork
with a full sized Unicode library for determining the actual length
of a given string on a terminal with a monospaced font -- such a
fork will be 4-5 times larger than Lunacy is right now.