Add support for <NOFMT> for non-LGC languages
So these tools can remain useful for non-LGC languages, we now have
a NOFMT tag which indicates we should not attempt to make lines short
enough to fit on a terminal screen line if a given UTF-8 code point
is not fixed width.

With such languages, hopefully the system `fmt` command will have
full Unicode support for the language in question.