Clarify fmt() code’s UTF-8 limitation
Each and every code point has to be the same width.  We can not
have any combining characters, double width characters, or zero
width characters in the UTF-8 input for the ej2man.new and ej2txt.new
tools for its internal fmt() routine to work.

Since the EJ tools have been around for 20 years and the underlying
docs have only been translated in to French (which, like English, keeps
the fixed-width assumption; even the uncommon œ and æ -- never used
in the Franch translation of the MaraDNS docs -- are squashed
to be only one character wide on monospace terminals), this is not an
issue in the real world.