EJ tools: Do not work with generic Lua
The thinking is this: MaraDNS is a stable, mature application.

That in mind, it will not be revised frequently.  Should Lua 5.5 or
whatever ever be revised such that it will not run the EJ tools, the
EJ tools should not stop running.  Instead, they will continue to
happily run with Lua 5.1 or a compatible fork.

This in mind, we only allow `lua5.1` or `lua-5.1`, which is a
common package (Both Ubuntu 22.04 and Alpine Linux 3.14 have it) or
lunacy (The Lua 5.1 fork included with MaraDNS).