Work on strictly POSIX-compliant makefiles for documents
Strict POSIX compliance requires the targets to be in the same
directory as the source.  Since the documents don’t do this, the
workaround is to have `Makefile.posix` make a bunch of symlinks,
which are then used as the deps for the rendered documents.

Here’s a pipeline to generate that part of the Makefile:

grep ../source/ Makefile | grep : | cut -f2 -d: | awk '
{for(a=1;a<=NF;a++){print $a}}' | awk -F/ '
{print $NF ":"; print "\tln -s " $0 " " $NF;print ""}' > foo

Since have all those symlinks is messy, I have a special
“make linkclean” rule in Makefile.posix which removes the links.