Move maraconf.sh to attic
Its #! line is /bin/bash, so I will assume it uses Bashisms, which go
against my goal to avoid non-standardized tools (or Lua5.1) in the
MaraDNS toolchain.

This script has not been updated or tested for well over 14 years;
the last update is from December 14, 2007.

If anyone wants it, it lingers at
https://github.com/samboy/MaraDNS-attic