Csv2_database.c: More work on Y2038 issues
Right now, we have a Y2841 issue with the synth SOA numbers.  It’s
4294 if we follow RFC1912.  2841 is far enough in to the future
that we can tell people using MaraDNS in the 2800s to be sure to
have a zone file with a specified SOA serial number in them, or we increase
the “second 1” epoch to be later than 1979.