Csv2_database.c: Correct handling of corner cases
If we can’t get a GMT timestamp for the file time, or if we’re
in MINGW32 *and* synth_soa_serial is 2 in the mararc file *and
also* the user doesn’t have a SOA with a serial number in the
zone file, there would be glitches.  Fixed.

The correct way to handle MINGW32 is to use the proprietary win32
API which will get a correct timestamp for a file well past the
Y2038 point in time, and use that to generate an appropriate SOA
serial number (if it’s not set in the zone file)