Y2038: Update mararc.ej to add Y2038 notes
timestamp_type and synth_soa_serial can not be set on systems
with a 32-bit time_t.  This is now in the documentation.

This is to avoid having MaraDNS use invalid time and date values starting
in late January of 2038; systems with a 32-bit time_t can very well
have their underlying system libraries with regards to dates and times
no longer correctly function come 2038.