Csv2_database: Y2038 compliance on Windows work
For this to be Y2038 compliant on Windows, I need to replace
stat() with Windows’s proprietary clunky API for getting file
information.  Win32 doesn’t have stat() for getting information
about an unopened file; it needs to open the file then get the
file time on the opened file.  Sigh.