blockListMake: Count how many bytes we need to store the strings
Here is the proposed format for a string chain:

<size><string> <size><string> 0xff

Here, size is a 2-byte big endian number
String is the string

0xff indicates a given list has ended

Each hash bucket gets its own list.

We also will have hash buckets.  This is a list of 32-bit big endian numbers
pointing to hash buckets.

Planned format:

\0DwB (Deadwood blocklist)
<Half SipHash key> (8 bytes)
<Hash size> (4 bytes, 32-bit big endian number)

Then an array of <hash size> 32-bit hash buckets.  Each of those
elements will point to a place in the combined string file.  If
the value is 0xffffffff, there is no string with the specified
hash bucket value.

Then we will have the combined string block, which is a bunch of string
chains.