Lunacy: Use editline, *not* readline
Since readline is GPL, and since MaraDNS is *not* GPL, better to
use the API-compatible editline library.  This way, MaraDNS users
can not accidentaly make a GPL binary by linking to readline.

Also: Lunacy in MaraDNS can be a desktop calculator, with support for
expressions which start with [0-9(] being run by Lunacy, without needing
a leading '='.  This is a one-line hack, but makes Lunacy a better
desktop calculator.  This feature is backported from the Git version
of lunacy.