coLunacyDNS and Lunacy: Sync Lunacy to Git version
The MaraDNS version of Lunacy has been out of sync with the Git
version of Lunacy for two years.  Resynchronize them; the reason
for the non-sync is because MaraDNS doesn’t need lfs nor spawner for
its use of Lunacy.

This in mind, I have added a note in the Lunacy Git tree to the
Lunacy documentation that it may not have lfs/spawner, added a
“Makefile.leaner” which compiles Lunacy w/o spawner and lfs, then
copied over the 2022-12-06 version of Lunacy to MaraDNS.

Because MaraDNS uses a different build process, and since there’s
no need for Spawner/Lfs in Mara, I have removed those libs and updated
the Makefiles to not use them at all.

Point being, MaraDNS and Git Lunacy are now mostly the same, if Git
Lunacy is compiled w/o Spawner/Lfs