blocked_hosts_hash_file dwood3rc parameter
The thinking is that Deadwood will, after entering the chroot_dir,
read this file to get a pre-made hash of host names to block.  This
allows a lot of host names to be blocked quickly while using little
memory, since the memory-heavy process of making the compact hash
will be done by a separate program.