Add support for wildcard searching with block hashes
For example, if porn.foo is in the block hash file, then we match
on porn.foo as well as bar.porn.foo, anything.bar.porn.foo, 12345.porn.foo,
and so on.

Likewise, if naughty.porn.foo is in the block hash file, then we match on
baz.naughty.porn.foo, bozzle.naughty.porn.foo, and so on.

The rule is this:

* Take the incoming query, like bozzle.naughty.porn.foo
* See if "bozzle.naughty.porn.foo" is in the block hash.  If so, block.
* Otherwise, convert "bozzle.naughty.porn.foo" in to "naughty.porn.foo",
  "porn.foo", and just "foo".  Look for all three in the database of
  blocked hosts.

Note that wildcards only work for domain names which are one, two, or
three labels long in the database.  If "really.naughty.porn.foo" is in
the database, "buzz.really.naughty.porn.foo" will *not* match.

As an aside, porn.bin needs to be in /etc/deadwood (or wherever the
chroot_dir value is) directory.